Refactor pfor tests into multiple files.
PiperOrigin-RevId: 228620325
